requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
langchain~=0.0.39
openai~=0.28.1
google-search-results ~= 2.4.2
flask~=3.0.0
pydantic~=1.10.13